requirements.txt
Flask
crewai
langgraph
langchain_community
langchain_google_genai
python-docx
PyMuPDF
